revamp signin component
