+++ client/src/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import Nav from './components/Nav';
-import SignUpForm from './components/SignupForm';
+import SignupForm from './components/SignupForm';
 import './App.css';
 
 function App() {
   return (
     <div>
       <Nav></Nav>
-      <SignUpForm></SignUpForm>
+      <SignupForm></SignupForm>
     </div>
   );
 }
